Update Python to 3.12 (latest)
Python 3.8 is a bit old at this point and a couple of dependencies
already require Python 3.10, making the current image incompatible.

One examples is https://github.com/AVATEAM-IT-SYSTEMHAUS/mkdocs-kroki-plugin/blob/main/setup.py#L20,
which my employer uses internally and thus we need a Python version
update to be able to use the latest version of that plugin.